+++ src/newlike.js
@@ -1,0 +1,15 @@
+const newLike = async (id) => {
+    const response = await fetch(
+      'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/IdQnbnyUh784FAUyhm2C/likes/',
+      {
+        method: 'POST',
+        body: JSON.stringify({
+          item_id: id,
+        }),
+        headers: { 'Content-Type': 'application/json' },
+      },
+    );
+    const like = await response.json();
+    return like;
+  };
+export default newLike;
